# documentation/openapi.yaml
openapi: 3.0.0
info:
  description: Node backend template API
  version: 1.0.0
  title: Node backend template
  contact:
    email: hola@houlak.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
  /users:
    post:
      tags:
        - user
      summary: Create user
      description: ''
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User to create
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
servers:
  - url: localhost:4040/api
components:
  schemas:
    User:
      type: object
      required:
        - name
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        mobileNumber:
          type: string
        deaf:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            type: string
        photo:
          type: string
          format: binary
        tyc:
          type: boolean
...